updated care_site_id from bigint to int

+++ source/org.ohdsi.cdm.framework.common/DataReaders/v5/LocationDataReader.cs
@@ -169,7 +169,7 @@ namespace org.ohdsi.cdm.framework.common.DataReaders.v5
             switch (i)
             {
                 case 0:
-                    return typeof(long);
+                    return typeof(int);
 
                 case 1:
                     return typeof(string);
